+++ Editor/src/com/kotcrab/vis/editor/module/scene/entitymanipulator/tool/RotateTool.java
@@ -21,23 +21,18 @@ import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer.ShapeType;
-import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.kotcrab.vis.editor.App;
-import com.kotcrab.vis.editor.module.editor.MouseLoopingModule;
-import com.kotcrab.vis.editor.module.scene.CameraModule;
 import com.kotcrab.vis.editor.proxy.EntityProxy;
+import com.kotcrab.vis.editor.util.gdx.ShapeRendererUtils;
 import com.kotcrab.vis.runtime.util.ImmutableArray;
 
 /** @author Kotcrab */
-public class RotateTool extends BaseGizmoTool {
+public class RotateTool extends AbstractGizmoTool {
 	public static final String TOOL_ID = App.PACKAGE + ".tools.RotateTool";
 
 	private static final Vector3 tmpV3 = new Vector3();
-
-	private CameraModule camera;
-	private MouseLoopingModule mouseLooping;
 
 	private float circleCenterX;
 	private float circleCenterY;
@@ -69,25 +64,14 @@ public class RotateTool extends BaseGizmoTool {
 			float startY = circleCenterY - innerRadius;
 			float endX = circleCenterX;
 			float endY = circleCenterY + innerRadius;
+
 			shapeRenderer.setColor(Color.RED);
-			drawLineRotatedAroundPoint(shapeRenderer, startX, startY, endX, endY, rotation, circleCenterX, circleCenterY);
+			ShapeRendererUtils.lineRotatedAroundPoint(shapeRenderer, startX, startY, endX, endY, rotation, circleCenterX, circleCenterY);
 			shapeRenderer.setColor(Color.GREEN);
-			drawLineRotatedAroundPoint(shapeRenderer, startX, startY, endX, endY, rotation + 90, circleCenterX, circleCenterY);
+			ShapeRendererUtils.lineRotatedAroundPoint(shapeRenderer, startX, startY, endX, endY, rotation + 90, circleCenterX, circleCenterY);
 
 			shapeRenderer.end();
 		}
-	}
-
-	private void drawLineRotatedAroundPoint (ShapeRenderer shapeRenderer, float startX, float startY, float endX, float endY, float angleDeg, float centerX, float centerY) {
-		//http://stackoverflow.com/a/12161405/1950897
-		float a = angleDeg * MathUtils.degreesToRadians;
-		float newStartX = centerX + (startX - centerX) * MathUtils.cos(a) - (startY - centerY) * MathUtils.sin(a);
-		float newStartY = centerY + (startX - centerX) * MathUtils.sin(a) + (startY - centerY) * MathUtils.cos(a);
-
-		float newEndX = centerX + (endX - centerX) * MathUtils.cos(a) - (endY - centerY) * MathUtils.sin(a);
-		float newEndY = centerY + (endX - centerX) * MathUtils.sin(a) + (endY - centerY) * MathUtils.cos(a);
-
-		shapeRenderer.line(newStartX, newStartY, newEndX, newEndY);
 	}
 
 	@Override
@@ -100,7 +84,6 @@ public class RotateTool extends BaseGizmoTool {
 		boolean insideOuterCircle = Math.pow(x - circleCenterX, 2) + Math.pow(y - circleCenterY, 2) <= Math.pow(outerRadius, 2);
 
 		insideRotateArea = outsideInnerCircle && insideOuterCircle;
-		if (insideRotateArea) mouseLooping.loopCursor();
 		return true;
 	}
 
@@ -108,6 +91,7 @@ public class RotateTool extends BaseGizmoTool {
 	public boolean touchDown (InputEvent event, float x, float y, int pointer, int button) {
 		boolean result = super.touchDown(event, x, y, pointer, button);
 		startingRotationValue = getRotationFromVirtualMouse();
+		if (insideRotateArea) mouseLooping.loopCursor();
 		return result;
 	}