+++ Editor/src/com/kotcrab/vis/editor/util/gdx/ShapeRendererUtils.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2014-2015 See AUTHORS file.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.kotcrab.vis.editor.util.gdx;
+
+import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
+import com.badlogic.gdx.math.Vector2;
+import com.kotcrab.vis.editor.util.VisMathUtils;
+
+/** @author Kotcrab */
+public class ShapeRendererUtils {
+	private static final Vector2 tmpV2 = new Vector2();
+
+	public static void lineRotatedAroundPoint (ShapeRenderer shapeRenderer, float startX, float startY, float endX, float endY, float angleDeg, float centerX, float centerY) {
+		VisMathUtils.rotatePointAroundCenter(startX, startY, tmpV2, centerX, centerY, angleDeg);
+		float newStartX = tmpV2.x;
+		float newStartY = tmpV2.y;
+
+		VisMathUtils.rotatePointAroundCenter(endX, endY, tmpV2, centerX, centerY, angleDeg);
+		float newEndX = tmpV2.x;
+		float newEndY = tmpV2.y;
+
+		shapeRenderer.line(newStartX, newStartY, newEndX, newEndY);
+	}
+}